mqtt_webapp/requirements.txt
Flask==2.0.1
paho-mqtt==1.6.1
python-dotenv==0.19.0
flask-socketio==5.1.1